refactor(inspection-service): extract inspection-report sub-service
Move getReportData (the resolved render-ready report aggregation),
getReportContentHash, and getReportPdfFooterContext into InspectionReportService.
getReportContentHash calls getReportData internally (same service). The facade
delegates getReportData so the analytics/annotations facade-callbacks still
resolve. Moved code is byte-identical (dynamic schema import re-pathed).
